+++ tests/genai/evaluate/test_utils.py
@@ -14,6 +14,7 @@ from mlflow.genai import scorer
 from mlflow.genai.evaluation.utils import (
     _convert_scorer_to_legacy_metric,
     _convert_to_eval_set,
+    validate_tags,
 )
 from mlflow.genai.scorers.builtin_scorers import RelevanceToQuery
 from mlflow.utils.spark_utils import is_spark_connect_mode
@@ -77,6 +78,7 @@ def sample_dict_data_multiple():
             "inputs": {"question": "What is MLflow?"},
             "outputs": "actual response for third question",
             "expectations": {},
+            "tags": {},
         },
     ]
 
@@ -466,3 +468,33 @@ def test_scorer_pass_through_aggregations(aggregations):
     legacy_metric_builtin = _convert_scorer_to_legacy_metric(builtin_scorer)
     assert legacy_metric_builtin.name == "relevance_to_query"
     assert legacy_metric_builtin.aggregations == builtin_scorer.aggregations
+
+
+@pytest.mark.parametrize(
+    "tags",
+    [
+        None,
+        {},
+        {"key": "value"},
+        {"env": "test", "model": "v1.0"},
+        {"key": 123},  # Values can be any type
+        {"key1": "value1", "key2": None},  # Values can be any type
+    ],
+)
+def test_validate_tags_valid(tags):
+    validate_tags(tags)
+
+
+@pytest.mark.parametrize(
+    ("tags", "expected_error"),
+    [
+        ("invalid", "Tags must be a dictionary, got str"),
+        (123, "Tags must be a dictionary, got int"),
+        ([1, 2, 3], "Tags must be a dictionary, got list"),
+        ({123: "value"}, "Tag keys must be strings, got int for key: 123"),
+        ({"key1": "value1", 123: "value2"}, "Tag keys must be strings, got int for key: 123"),
+    ],
+)
+def test_validate_tags_invalid(tags, expected_error):
+    with pytest.raises(MlflowException, match=expected_error):
+        validate_tags(tags)
